chore: reorganize component
